+++ src/pages/catalog/components/catalog-item/catalog-item.tsx
