+++ src/pages/catalog/components/catalogFilter/burger-filter.tsx
@@ -1,6 +1,6 @@
 import { ReactElement, useEffect, useRef } from 'react';
 import { ProductProjectionResponse } from '../../../../api/catalog/products.types.ts';
-import { CatalogFilter, Filters } from './CatalogFilter.tsx';
+import { CatalogFilter, Filters } from './catalog-filter.tsx';
 
 export function BurgerFilter({
   showBurger,
