+++ src/pages/catalog/components/catalogFilter/components/get-filter-reset-svg.tsx
