+++ src/pages/catalog/components/category-bar/Category-bar.tsx
